requirements.txt
fastapi
uvicorn
python-dotenv
openai
langchain
langchain-community
langchain-huggingface
langchain-text-splitters
faiss-cpu
pypdf
sentence-transformers